Improve `RSN` retrieval
This commit updates `retrieve_peer_gm_data` so that it does not rely
on a particular groupping of the PEER ground motion files, but instead
dynamically searches all available directories until it identifies the
record. This makes it much easier to expand the available ground
motions by simply placing more folders in the directory.